Organise consensus into modules
