+++ consensus/src/lib.rs
@@ -1,16 +1,12 @@
 // Copyright(C) Facebook, Inc. and its affiliates.
 #[cfg(feature = "dolphin")]
-mod committer;
-#[cfg(feature = "dolphin")]
 mod dolphin;
 mod state;
 #[cfg(not(feature = "dolphin"))]
 mod tusk;
-#[cfg(feature = "dolphin")]
-mod virtual_state;
 
 #[cfg(not(feature = "dolphin"))]
 pub use crate::tusk::Tusk;
 
 #[cfg(feature = "dolphin")]
-pub use crate::dolphin::Dolphin;
+pub use crate::dolphin::dolphin::Dolphin;
